backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
transformers>=4.35.2
torch>=2.6.0
torchvision>=0.19.0
Pillow>=10.1.0
numpy>=1.24.3
requests>=2.31.0
aiofiles>=23.2.1
google-generativeai>=0.3.0
